requirements-lock.txt
# This file is auto-generated by update-requirements.sh
# DO NOT EDIT MANUALLY - Use scripts/update-requirements.sh instead
# Generated on: $(date)

# Core dependencies with exact versions
fastapi==0.118.0
uvicorn[standard]==0.30.6
pydantic==2.6.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.32.5

# Add other dependencies as needed
# This file will be populated by the update script
